server/requirements.txt
aiodns==1.1.1
aiofiles==0.3.2
aiohttp==2.3.2
AoikLiveReload==0.1.0
argh==0.26.2
astroid==1.5.3
async-timeout==2.0.0
asyncpg==0.13.0
cchardet==2.1.1
chardet==3.0.4
click==6.7
gunicorn==19.7.1
httptools==0.0.9
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==3.3.2
olefile==0.44
pathtools==0.1.2
Pillow==4.3.0
ptvsd==3.2.1
pycares==2.3.0
pylint==1.7.4
python-dotenv==0.7.1
PyYAML==3.12
sanic==0.6.0
six==1.11.0
ujson==1.35
uvloop==0.8.1
watchdog==0.8.3
websockets==4.0.1
Willow==1.1
wrapt==1.10.11
yarl==0.13.0
